+++ RiskOfChaos/Config/Configs.cs
@@ -1,5 +1,6 @@
 ﻿using BepInEx.Configuration;
 using RiskOfOptions;
+using UnityEngine;
 
 namespace RiskOfChaos
 {
@@ -7,6 +8,21 @@ namespace RiskOfChaos
     {
         const string CONFIG_GUID = $"RoC_Config_General";
         const string CONFIG_NAME = $"Risk of Chaos: General";
+
+        public static readonly Sprite GenericIcon;
+
+        static Configs()
+        {
+            Texture2D genericIconTexture = new Texture2D(256, 256);
+            if (genericIconTexture.LoadImage(Properties.Resources.icon))
+            {
+                GenericIcon = Sprite.Create(genericIconTexture, new Rect(0f, 0f, genericIconTexture.width, genericIconTexture.height), new Vector2(0.5f, 0.5f));
+            }
+            else
+            {
+                Log.Error("Failed to load config icon");
+            }
+        }
 
         internal static void Init(ConfigFile file)
         {
@@ -20,7 +36,11 @@ namespace RiskOfChaos
             Debug.Bind(file);
 #endif
 
-            // ModSettingsManager.SetModIcon(general_icon, GENERAL_GUID, GENERAL_NAME);
+            if (GenericIcon)
+            {
+                ModSettingsManager.SetModIcon(GenericIcon, CONFIG_GUID, CONFIG_NAME);
+            }
+
             ModSettingsManager.SetModDescription("General config options for Risk of Chaos", CONFIG_GUID, CONFIG_NAME);
         }
     }