+++ RiskOfChaos/Patches/ProjectileSpeedHook.cs
@@ -70,7 +70,14 @@ namespace RiskOfChaos.Patches
             {
                 new Hook(orbDurationSetter, (Action<Orb, float> orig, Orb self, float value) =>
                 {
-                    tryMultiplyProjectileTravelTime(ref value, self);
+#pragma warning disable Publicizer001 // Accessing a member that was not originally public
+                    float distanceToTarget = self.distanceToTarget;
+#pragma warning restore Publicizer001 // Accessing a member that was not originally public
+
+                    if (distanceToTarget > 0f)
+                    {
+                        tryMultiplyProjectileTravelTime(ref value, self);
+                    }
 
                     orig(self, value);
                 });